+++ client/src/index.tsx
@@ -4,8 +4,11 @@ import { Router } from "@solidjs/router";
 import "virtual:uno.css";
 import "@unocss/reset/tailwind.css";
 import "./styles/global.css";
+import "./styles/themes-structure.css";
 import "./styles/themes.css";
+import "./styles/themes-pixel.css";
 import "./styles/highlight-theme.css";
+import "./styles/themes-pixel-highlight.css";
 import AppRoutes from "./App";
 
 const root = document.getElementById("root");